refactor: renamed helpers package to keyspace
