+++ provider/internal/keyspace/key.go
@@ -1,4 +1,4 @@
-package helpers
+package keyspace
 
 import (
 	"crypto/sha256"
