+++ provider/internal/keyspace/key_test.go
@@ -1,4 +1,4 @@
-package helpers
+package keyspace
 
 import (
 	"crypto/rand"
