+++ provider/internal/keyspace/trie.go
@@ -1,4 +1,4 @@
-package helpers
+package keyspace
 
 import (
 	"github.com/libp2p/go-libp2p/core/peer"
